requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
langchain-core
langchain-google-genai
langgraph
qdrant-client
